Modified 'won' attribute in user_game_attempts to be proper length for boolean value

+++ migrations/1-initial.js
@@ -849,7 +849,7 @@ var migrationCommands = [{
                     "allowNull": true
                 },
                 "won": {
-                    "type": Sequelize.INTEGER(4),
+                    "type": Sequelize.INTEGER(1),
                     "allowNull": true
                 },
                 "created_at": {
